feat(#94): refactor tsconfig and add type-checking to linting pipeline fix(#94): remove `isError` from `useLocalStorage` fix(#94): add types file for `?raw` import
