+++ app/.storybook/preview/preview.tsx
@@ -1,8 +1,8 @@
 import { HocComposer, RoutingService } from '@ds/core'
 import { argTypesEnhancers, getDocsConfig, getStoryConfig, mockNavigate, toolbarConfig } from '@ds/docs/core'
 import { type Preview } from '@storybook/nextjs-vite'
-import { initClientI18n } from '../src/core/i18n/i18n-client'
-import { I18nProvider } from '../src/core/i18n/i18n-provider'
+import { initClientI18n } from '../../src/core/i18n/i18n-client'
+import { I18nProvider } from '../../src/core/i18n/i18n-provider'
 import './styles'
 
 LOG('BUILD_NUMBER:', ENV__BUILD_NUMBER)
