+++ app/.storybook/preview/types.d.ts
@@ -1,4 +1,5 @@
 import '@ds/docs/types.d.ts'
+import '@ds/types.d.ts'
 
 declare global {
   var ENV__BASE_PATH: string
