--- app/.storybook/styles/tailwind.config.ts
@@ -1,12 +1,0 @@
-import { type Config } from 'tailwindcss'
-import { tailwindTheme } from '../../../ds/dist/scripts/tailwind'
-
-const config: Config = {
-  content: [
-    '../../{src}/**/*.{ts,tsx,mdx}', // app + docs
-    '../../../ds/{src,docs}/**/*.{ts,tsx,mdx}', // ds + docs
-  ],
-  theme: tailwindTheme,
-}
-
-export default config
