+++ app/tsconfig-sb.json
@@ -18,5 +18,5 @@
       "@ds/*": ["../ds/dist/*"]
     }
   },
-  "include": [".storybook/preview.tsx", ".storybook/preview.types.d.ts", "src/**/*.stories.ts", "src/**/*.stories.tsx"]
+  "include": [".storybook/preview/**/*", "src/**/*.stories.ts", "src/**/*.stories.tsx", "src/**/*.mdx"]
 }
