+++ app/vitest.config.ts
@@ -20,7 +20,7 @@ export default defineConfig({
             headless: true,
             instances: [{ browser: 'chromium' }],
           },
-          setupFiles: ['./.storybook/vitest.setup.ts'],
+          setupFiles: ['./.storybook/preview/vitest.setup.ts'],
         },
       },
     ],
