+++ ds/.storybook/preview/styles/tailwind.config.ts
@@ -1,9 +1,9 @@
 import { type Config } from 'tailwindcss'
-import { tailwindTheme } from '../dist/scripts/tailwind'
+import { tailwindTheme } from '../../../dist/scripts/tailwind'
 
 const config: Config = {
   content: [
-    '../{src,docs}/**/*.{ts,tsx,mdx}', // ds + docs
+    '../../../{src,docs}/**/*.{ts,tsx,mdx}', // ds + docs
   ],
   theme: tailwindTheme,
 }
