+++ ds/docs/stories/2-get-started.mdx
@@ -195,7 +195,7 @@ You can also generate a build number based on build time.
 Add TypeScript type definitions for the documentation components:
 
 <DocsCodeBlock
-  filename=".storybook/preview.types.d.ts"
+  filename=".storybook/preview/types.d.ts"
   code={`
 		import '@ds/docs/types.d.ts'
 	`}
